[bugfix] Fix setup.read_desc() regex
If a space was behind a sphinx link and the link was different from
the label, the link was removed but the alignment does not match the
previous line. Modify the regex pattern to remove a following space
in such case.

Also test the long_description created by setup.py in make_dist.py
script

Bug: T327056

+++ make_dist.py
@@ -45,7 +45,7 @@ import abc
 import shutil
 import sys
 from pathlib import Path
-from subprocess import check_call
+from subprocess import check_call, run
 
 import setup
 from pywikibot import __version__, error, info, input_yn, warning
@@ -108,6 +108,10 @@ class SetupBase(abc.ABC):
             return
         finally:
             self.cleanup()
+
+        # check description
+        if run('twine check dist/*', shell=True).returncode:
+            return
 
         if self.local:
             check_call('pip uninstall pywikibot -y', shell=True)
